requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
iniconfig==2.0.0
install==1.3.5
packaging==23.2
playwright==1.40.0
PyHamcrest==2.1.0
pytest==7.4.4
pytest-base-url==2.0.0
pytest-playwright==0.4.3
requests==2.31.0
urllib3==2.1.0
allure-pytest==2.9.41
pyyaml==6.0.1
